renamed override-stream-write to overrideStreamWrite to be more consistent with the rest of the file names
