+++ loggers/overrideStreamWrite.js
